[5주차][기본] 서비스 인터페이스, 구현체 분리
- 기존존 서비스 코드 계층 나누기
- 의존성 역전 적용: Interface → 구현체 주입
- 각 계층 명확히 구분하여 리팩토링
