+++ src/main/java/kr/hhplus/be/server/src/application/service/ConcertSeatServiceImpl.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.src.application.service;
+
+public class ConcertSeatServiceImpl {
+
+
+}
