+++ src/main/java/kr/hhplus/be/server/src/application/service/SeatServiceImpl.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.src.application.service;
+
+public class SeatServiceImpl {
+}
