+++ src/main/java/kr/hhplus/be/server/src/interfaces/booking/BookingController.java
@@ -4,12 +4,10 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import kr.hhplus.be.server.src.common.ResponseMessage;
 import kr.hhplus.be.server.src.domain.booking.Booking;
-import kr.hhplus.be.server.src.interfaces.payment.PaymentResponse;
 import kr.hhplus.be.server.src.domain.booking.BookingService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
-//콘서트 예약 기능
 @Tag(name = "예약", description = "예약 관련 API")
 @RestController
 @RequiredArgsConstructor
@@ -17,7 +15,6 @@ import org.springframework.web.bind.annotation.*;
 public class BookingController {
 
     private final BookingService bookingService;
-//    private final BookingPaymentFacade bookingPaymentFacade;
 
     @Operation(summary = "좌석 예약 요청", description = "좌석 예약을 요청합니다.")
     @PostMapping("/seats")
